+++ src/Peer.cpp
@@ -336,7 +336,7 @@ void P::UDP::threadFunc() {
 				case Code::Success: {
 					if (pack.type() == Type::Ping) {
 						Message::Ping ping;
-						if (pack(ping, packet.size() - sizeof(NetHeader))) {
+						if (pack(ping, static_cast< uint32_t >(packet.size()) - sizeof(NetHeader))) {
 							if (m_feedback.ping) {
 								m_feedback.ping(endpoint, ping);
 							}
